Update index.html
Added to the  class 'text', -h1, -h2, -h3  in the simulated code, to style those sentences approprietly.
Added some missing >/span> elements in the code blocks code. It wasn't causing any visible issues currently, but it was creating spans that enclosed more than it should. That would most definitely cause issues in future styling, if it was left as it was.